refactor(dashboard): Reworked the chest count per account query to show all accounts
