+++ apps/website/src/actions/chest.ts
@@ -46,19 +46,3 @@ export const getAllChests = async () => {
     accountCount: countQuery.accountCount,
   };
 };
-
-export const getChestCountPerAccount = async () => {
-  const result = await db
-    .selectFrom("chest")
-    .innerJoin("account", "chest.accountId", "account.id")
-    .select((eb) => [
-      "account.name as account",
-      "account.townhall as townhall",
-      eb.fn.count<number>("chest.id").as("count"),
-    ])
-    .groupBy(["chest.accountId", "account.name", "account.townhall"])
-    .orderBy("account.townhall", "desc")
-    .execute();
-
-  return result;
-};
